requirements/base.txt
Django==1.10.2
django-cors-headers==2.0.0
django-mptt==0.8.6
djangorestframework==3.5.3
Pillow==4.0.0
requests==2.13.0
